backend/requirements.txt
Django==5.1.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
Pillow
